requirements.txt
Django==4.2.7
Pillow==10.0.1
google-generativeai==0.3.2
python-decouple==3.8
djangorestframework==3.14.0
requests==2.31.0
